.......... [ZBXNEXT-2662] updated to the latest trunk (without conflicts)
git-svn-id: svn://svn.zabbix.com/branches/dev@54523 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-2662-6/src/libs/zbxsysinfo/common/file.c
@@ -246,11 +246,11 @@ err:
 
 int	VFS_FILE_REGEXP(AGENT_REQUEST *request, AGENT_RESULT *result)
 {
-	char	*filename, *regexp, encoding[32], *output, *start_line_str, *end_line_str;
-	char	buf[MAX_BUFFER_LEN], *utf8, *tmp, *ptr;
-	int	nbytes, f = -1, ret = SYSINFO_RET_FAIL;
-	size_t	start_line, end_line, current_line = 0;
-	double	ts;
+	char		*filename, *regexp, encoding[32], *output, *start_line_str, *end_line_str;
+	char		buf[MAX_BUFFER_LEN], *utf8, *tmp, *ptr;
+	int		nbytes, f = -1, ret = SYSINFO_RET_FAIL;
+	uint32_t	start_line, end_line, current_line = 0;
+	double		ts;
 
 	ts = zbx_time();
 
@@ -367,11 +367,11 @@ err:
 
 int	VFS_FILE_REGMATCH(AGENT_REQUEST *request, AGENT_RESULT *result)
 {
-	char	*filename, *regexp, *tmp, encoding[32];
-	char	buf[MAX_BUFFER_LEN], *utf8, *start_line_str, *end_line_str;
-	int	nbytes, len, res, f = -1, ret = SYSINFO_RET_FAIL;
-	size_t	start_line, end_line, current_line = 0;
-	double	ts;
+	char		*filename, *regexp, *tmp, encoding[32];
+	char		buf[MAX_BUFFER_LEN], *utf8, *start_line_str, *end_line_str;
+	int		nbytes, len, res, f = -1, ret = SYSINFO_RET_FAIL;
+	uint32_t	start_line, end_line, current_line = 0;
+	double		ts;
 
 	ts = zbx_time();
